Fix overflow in genome shrink
This was causing infinite loops during certain large alignments.
`std::string::size` returns a `size_t`, which is 8 bytes wide on
64-bit Linux but `int` is only 4 bytes wide (and signed, to boot).

+++ src/SQLiteConstructor.cpp
@@ -1655,12 +1655,12 @@ void SQLiteConstructor::clean_shrunken_genomes() {
 	string trimmed_genomes = genefoldername + "genome_shrink_aln";
 	fu.read_aligned_fasta_into(tempalseqs, trimmed_genomes); // assuming this is aligned
 
-	int seqlength = tempalseqs[0].get_sequence().size();
+	size_t seqlength = tempalseqs[0].get_sequence().size();
 	float fseql = float(tempalseqs.size());
-	vector<int> removeem;
-	for (int j = 0; j < seqlength; j++) {
-		int gaps = 0;
-		for (int i = 0; i < tempalseqs.size(); i++) {
+	vector<size_t> removeem;
+	for (size_t j = 0; j < seqlength; j++) {
+		size_t gaps = 0;
+		for (size_t i = 0; i < tempalseqs.size(); i++) {
 			if (tempalseqs[i].get_sequence()[j] == '-' || tempalseqs[i].get_sequence()[j] == 'N' || tempalseqs[i].get_sequence()[j] == 'n')
 				gaps += 1;
 		}
@@ -1669,9 +1669,9 @@ void SQLiteConstructor::clean_shrunken_genomes() {
 			removeem.push_back(j);
 		}
 	}
-	for (int i = 0; i < tempalseqs.size(); i++) {
+	for (size_t i = 0; i < tempalseqs.size(); i++) {
 		string a;
-		for (int j = 0; j < seqlength; j++) {
+		for (size_t j = 0; j < seqlength; j++) {
 			if (count(removeem.begin(), removeem.end(), j) == 0)
 				a += tempalseqs[i].get_sequence()[j];
 		}
